Use import_tasks and fix ssh-rsa config  (#10)
PubkeyAcceptedAlgorithms was set in the config and it was introduced
in OpenSSH version 8.5. This is incompatible with Debian 9-11 as they
have older versions in their package repositories. Furthermore, ssh-rsa
is enabled by default until version 8.8 so setting it is only needed
for Debian versions above 12. This commit changes the configuration
to only include the setting for those Debian versions. Additionally
the deprecated "import" was changed to "import_tasks".